Moved forms into separate python modules
